FINAL PDF COMBINED
It was unsure if I needed to upload every single document I work with as it was suggested by part 4, so I end up doing just that just in case. I tried to make it as clear and organised as possible in GitHub, but I am still bit green here. Something to note is that for some reason the molecule labelled as Temcen did not run to full completion on archie with 2 cpus, unsure as to why even though I tried to troubleshoot it. I ended up just doing from 4 cpus, 8, cpus and 16 cpus used as to illustrate anyway the purpose of the excercise. I included the attempt at 2 cpus for both as well.